move and fix some tests
