+++ test/model_context_protocol/shared/configuration_test.rb
