+++ test/model_context_protocol/shared/instrumentation_test.rb
