+++ test/model_context_protocol/shared/methods_test.rb
