+++ test/model_context_protocol/shared/prompt_test.rb
