+++ test/model_context_protocol/shared/string_utils_test.rb
