+++ test/model_context_protocol/shared/tool/input_schema_test.rb
