+++ test/model_context_protocol/shared/tool_test.rb
